pyproject.toml
[project]
name = "BankOperations"
version = "0.1.0"
description = ""
authors = [
    {name = "Vic766",email = "766@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]

[tool.poetry]
packages = [{include = "bankoperations", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
src = [
    "mypy (>=1.19.1,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
tests = [
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
